requirements.txt
aiogram==3.2.0
python-dotenv==1.0.0
tortoise-orm==0.20.0
asyncpg==0.29.0
python-telegram-bot==20.7
